backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
ortools==9.8.3296
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
geopy==2.4.1
requests==2.31.0
python-dotenv==1.0.1
google-generativeai==0.3.2
httpx==0.26.0
